axolotl: keep using OMX
We are in a weird situation, where we are a 64 bit only device
with support for OMX (even if this is hacky).

In future we need to move away from OMX to C2 or revert back
to a 32+64 configuration, but for now to not break this branch
keep the status quo.

Reference: https://review.lineageos.org/399102

Change-Id: I8af292800ce7cc1f0c4940b7fb1980d9712d7b1d
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>